Refactor into library crate
