+++ Cargo.toml
@@ -4,6 +4,10 @@ version = "0.1.0"
 edition = "2024"
 license = "ISC"
 
+[lib]
+name = "vk"
+path = "src/lib.rs"
+
 [dependencies]
 clap = { version = "4.5.3", features = ["derive"] }
 reqwest = { version = "0.11.23", features = ["json", "rustls-tls"] }
