+++ docs/vk-design.md
@@ -41,9 +41,9 @@ completion.
 Runtime flags and subcommand options live in
 [src/cli_args.rs](../src/cli_args.rs). Keeping these structures in a dedicated
 module isolates the lint expectations generated by `clap` and keeps `main.rs`
-focused on orchestrating API calls and printing results. The public
-`GlobalArgs`, `PrArgs`, and `IssueArgs` structures are fully documented so
-their purpose and merge semantics are clear to downstream users.
+focused on parsing arguments and invoking the library. The public `GlobalArgs`,
+`PrArgs`, and `IssueArgs` structures are fully documented so their purpose and
+merge semantics are clear to downstream users.
 
 ## Utility
