+++ src/cli_args.rs
@@ -3,7 +3,6 @@
 //! Isolates clap derivations so lint expectations remain scoped, keeping
 //! `main.rs` focused on runtime logic.
 #![expect(non_snake_case, reason = "clap generates non-snake-case modules")]
-#![expect(unused_imports, reason = "clap derives import the struct internally")]
 
 use clap::Parser;
 use ortho_config::OrthoConfig;
